Switch to using system fonts
The changes in the system fonts are taken from StackOverflow's blog.
Since we can't set up Cache-Control headers when using Github Pages, the
fonts aren't cached.
The first paint of the website (with the system fonts) and the
subsequent renders (once the non-system fonts have been fetched) are
very different, making the overall site look un-snappy.

If Github ever adds the use of custom HTTP headers for Github Pages,
this change should be reverted.

Issue to track for the same:
    https://github.com/community/community/discussions/11884

Signed-off-by: Gunwant Jain <mail@wantguns.dev>